requirements.txt
appdirs==1.4.4
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
black==19.10b0
certifi==2020.6.20
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
dataclasses==0.7
dlib==19.6.1
flake8 @ file:///tmp/build/80754af9/flake8_1601911421857/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
imutils==0.5.3
kiwisolver @ file:///C:/ci/kiwisolver_1603996656077/work
matplotlib @ file:///C:/ci/matplotlib-base_1603357981283/work
mccabe==0.6.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
olefile==0.46
opencv-python==4.4.0.46
pathspec==0.7.0
Pillow @ file:///C:/ci/pillow_1603822370986/work
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex @ file:///C:/ci/regex_1602770568056/work
scipy @ file:///C:/ci/scipy_1597675683670/work
six==1.15.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
